Update ch01-03-hello-cargo.md
Remove field `[author]` as it is not added any more by default. Go removed in version 1.53.0 as it caused more problems than it solved. 

See here for the original English version of the book https://github.com/rust-lang/book/issues/2764
and this discussion on Stack Overflow https://stackoverflow.com/a/67891977